TST: test_monte_carlo.py reorganized
